NO-SPE: Fixing compilation warnings

+++ src/libpgcode/src/Slic3r/Biz/libpgcode/ProcessorImpl.cpp
@@ -124,11 +124,11 @@ ProcessorResult ProcessorImpl::finalize()
             }
         }
         // Now move all the vertices into place:
-        int orig_idx = moves.size() - 1;
+        int orig_idx = int(moves.size()) - 1;
         moves.resize(moves.size() + mod_moves.size());
         for (int i = int(moves.size()) - 1; i >= 0; --i) {
             moves[i] = moves[orig_idx];
-            if (! mod_moves.empty() && mod_moves.back() == orig_idx) {
+            if (! mod_moves.empty() && static_cast<int>(mod_moves.back()) == orig_idx) {
                 // What we just copied should be preceded by its copy (except for some fields).
                 moves[i-1] = moves[i];
                 MoveVertex& v = moves[i-1];
