+++ src/slic3r-shared/src/Slic3r/App/MaterialSettingsDialog.cpp
@@ -141,7 +141,7 @@ void MaterialSettingsDialog::on_tab_selected(int current_index)
 {
     AbstractSettingsDialog::on_tab_selected(current_index);
 
-    if (m_current_tab && current_index < m_material_cbi_list.size()) {
+    if (m_current_tab && current_index < static_cast<int>(m_material_cbi_list.size())) {
         m_current_preset_label->set_current_list(current_index);
     }
 }
